requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.94.1
h11==0.14.0
httpcore
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.6
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.3
